backend/requirements.txt
# FastAPI 后端依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LangChain 和 LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20
langgraph-checkpoint-sqlite>=1.0.0
aiosqlite>=0.19.0

# 文档处理
pypdf>=3.17.0
python-docx>=1.1.0

# 向量数据库
faiss-cpu>=1.7.4

# OpenAI
openai>=1.3.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 日志和监控
structlog>=23.2.0

# 开发工具 (可选)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0